refactoring:make textinput reusable, moving styles to common that are common across app
