+++ src/components/styles/AllDone.scss
@@ -3,15 +3,3 @@
     padding-top: 1%;
     color:var(--black);
 }
-#contactSales{
-  color: #fff;
-  background-color: #3f51b5;
-  /* margin: 10%; */
-  text-align: center;
-  /* float: right; */
-  margin-left: 12%;
-  margin-bottom: 10%;
-  width: 80%;
-  margin-top: 10%;
-  border-radius: 4rem;
-}
