--- src/components/styles/DeliverySchedule.scss
@@ -1,52 +1,0 @@
-/* article posting form container */
-.about-user-form-container {
-    display:flex;
-    flex-direction:column;
-    justify-content:flex-start;
-    transition:flex 0.3s ease-out; 
-    border-left: 0.5rem solid #968989;
-    box-shadow: grey 0.1rem 0.05rem 0.1rem;
-    flex:1;
-    width: 90%;
-    margin-left: 2%;
-    padding: 5px;
-    background-color:var(--white);
-  }
-  
-  
-  
-@media only screen and (min-width:600px) {
-    .BtnGroup {
-      margin-top: 5%;
-      overflow: hidden;
-      width: 100%;
-      margin-bottom: 10%;
-      
-}
-
-.BtnGroup :first-child {
-    float: none;
-    margin-left: 5%;
-    
-}
-
-.BtnGroup :nth-child(2) {
-    margin-right: 5%;
-    float: right;
-    
-}
-
-
-#next-with-previous,#previous  {
- 
-  padding: 10px 25px;
-  border: none;
-  color: #fff;
-  cursor: pointer;
-  border-radius: 4rem;
-  outline: none;
-  width: 30%;
-}
-
-    
-  }
