+++ src/containers/AboutUser.jsx
@@ -2,11 +2,11 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { createUser } from '../store/actions/signup-action';
 import { saveUser } from '../store/actions/users-action';
-import Button from '@material-ui/core/Button';
 import TextField from '@material-ui/core/TextField';
 import MenuItem from '@material-ui/core/MenuItem';
-import './styles/AboutUser.scss';
 import { roles,restaurantTypes} from '../data/Constants';
+import ButtonsGroup from '../components/ButtonsGroup';
+import FormHeader from '../components/FormHeader';
 
 class AboutUser extends Component {
   constructor(props) {
@@ -37,7 +37,6 @@ class AboutUser extends Component {
     console.log('data with next click',data);
     this.props.goNext();
     this.props.createUser(data);
-    // this.props.onSubmit(data);
   };
 
 
@@ -55,10 +54,8 @@ class AboutUser extends Component {
 
   render() {
     return (
-      <form  className="registration-form" onSubmit={this.onSubmit} autoComplete="off">
-        <div className="form-header">
-          <h1>About You</h1>
-        </div>
+      <form  className="signup-form" onSubmit={this.onSubmit} autoComplete="off">
+        <FormHeader headerText="About You"/>
         <div className="flex-wrap">
           <TextField
             required
@@ -154,27 +151,7 @@ class AboutUser extends Component {
             onChange={this.onChange}
           />
           <br/>
-          <div className="BtnGroup">
-            <Button
-              label="PREVIOUS"
-              onClick={this.goPrevious}
-              id="previous"
-              type="button"
-              variant="contained"
-              color="primary"
-            >
-            PREVIOUS
-            </Button>
-            <Button
-              label="NEXT"
-              id="next-with-previous"
-              type="submit"
-              variant="contained"
-              color="primary"
-            >
-            NEXT
-            </Button>
-          </div>
+          <ButtonsGroup goPrevious={this.goPrevious}/>
         </div>
       </form>
     );
